650 - filestore message length fix
